Separate the user tests
