+++ users/tests/test_registration.py
@@ -39,36 +39,3 @@ class RegistrationTest(TestCase):
         create_user('User1', 'User1@gmail.com', 'isp123456')
         response = self.client.post(reverse('register'), {'username':'User1','email':'User6@gmail.com','password1':'isp123456','password2':'isp123456'})
         self.assertEqual(response.status_code, 200)
-
-class LogInTest(TestCase):
-    """Class for test login system"""
-
-    def test_simple_login(self):
-        """If can login, that user must be authenticated in website"""
-        create_user("User1", "User1@gmail.com", "isp123456")
-        response = self.client.post(reverse('login'), {'username':'User1', 'password':'isp123456'}, follow=True)
-        self.assertTrue(response.context['user'].is_authenticated)
-
-    def test_can_not_login_with_user_that_not_enter_password(self):
-        """Can not login with enter the username but not enter the password"""
-        create_user("User1", "User1@gmail.com", "isp123456")
-        response = response = self.client.post(reverse('login'), {'username':'User1'}, follow=True)
-        self.assertFalse(response.context['user'].is_authenticated)
-
-    def test_can_not_login_with_user_that_enter_wrong_password(self):
-        """Can not login, if enter the wrong password"""
-        create_user("User1", "User1@gmail.com", "isp123456")
-        response = self.client.post(reverse('login'), {'username':'User1', 'password':'isp555555'}, follow=True)
-        self.assertFalse(response.context['user'].is_authenticated)
-
-class LogOutTest(TestCase):
-    """Class for test logout system"""
-
-    def test_login_and_logout(self):
-        """If you still login is_authenticated is true, if logout s_authenticated must be false"""
-        create_user("User1", "User1@gmail.com", "isp123456")
-        response = self.client.post(reverse('login'), {'username':'User1', 'password':'isp123456'}, follow=True)
-        self.assertTrue(response.context['user'].is_authenticated)
-        response = self.client.get(reverse('logout'))
-        self.assertFalse(response.context['user'].is_authenticated)
-
